requirements.txt
torch
tqdm
requests
python-dotenv